Produce a separate log message before each EOL.
This commit changes the behaviour of `logging.writer` to produce a
separate log message before each end-of-line (EOL) is encountered in the
string.

The writer also internally buffers any strings not terminated with an EOL
character until the next call to `Write()` will have string with an EOL character in it.
Then the entire buffered content before an EOL character is produced into a
separate log message.